backend/requirements.txt
web3
flask
flask-restplus
py-solc
flask-restful
flask-marshmallow
solc-select
Werkzeug==0.16.1